moved the stringutils out of ApiBukkit and into libMinecraft

+++ ApiBukkit/src/main/java/de/cubeisland/ApiBukkit/ApiConfiguration.java
@@ -14,7 +14,7 @@ import org.bukkit.configuration.ConfigurationSection;
 public class ApiConfiguration
 {
     public final ApiLogLevel logLevel;
-    public final int port;
+    public final short port;
     public final String authKey;
     public final int maxContentLength;
     public final Map<String, Collection<String>> disabledActions;
@@ -35,7 +35,7 @@ public class ApiConfiguration
             ApiBukkit.logException(e);
         }
         this.logLevel = tmpLogLevel;
-        this.port = config.getInt("Network.port");
+        this.port = (short)config.getInt("Network.port");
         this.authKey = config.getString("Network.authKey");
         this.maxContentLength = config.getInt("Network.maxContentLength");
